requirements.txt
openai
langchain
PyPDF2
docx2txt
fastapi
uvicorn
tiktoken
python-dotenv
faiss-cpu
jinja2
python-multipart
sqlalchemy
pandas